requirements.txt
redis==5.0.1
requests==2.32.4
python-dotenv==1.0.1
websockets==12.0
protobuf==4.25.3
